// apps/frontend/modules/finances/types.ts
export interface AnalyticSeriesRow {
  dimensions: Array<{
    name: 'budget'
    path: string
  }>
  metric: AnalyticMetric
  unit: 'DAI'
  value: number
  sum: number
}

export interface AnalyticSeries {
  period: string
  start: string
  end: string
  rows: AnalyticSeriesRow[]
}

export interface Analytic {
  series: AnalyticSeries[]
}

export interface BudgetMetric {
  actuals: ValueAndUnit
  budget: ValueAndUnit
  forecast: ValueAndUnit
  paymentsOnChain: ValueAndUnit
  paymentsOffChainIncluded: ValueAndUnit
  protocolNetOutflow: ValueAndUnit
}

export type BreakdownBudgetAnalytic = Record<string, BudgetMetric[]>

export interface Budget {
  parentId: string | null
  name: string
  image: string
  idPath: string
  id: string
  description: string
  codePath: string
  code: string
}

export type AnalyticMetric =
  | 'Budget'
  | 'Forecast'
  | 'PaymentsOnChain'
  | 'ProtocolNetOutflow'
  | 'Actuals'